requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# FastAPI + ASGI
# ─────────────────────────────────────────────────────────────────────────────
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# ─────────────────────────────────────────────────────────────────────────────
# Pydantic / Settings
# ─────────────────────────────────────────────────────────────────────────────
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ─────────────────────────────────────────────────────────────────────────────
# PostgreSQL + SQLAlchemy (async)
# ─────────────────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.23
asyncpg>=0.29.0
# ─────────────────────────────────────────────────────────────────────────────
# JWT + Passwords (bcrypt напрямую; passlib не используется из-за несовместимости с bcrypt 4.1+)
# ─────────────────────────────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# ─────────────────────────────────────────────────────────────────────────────
# Telegram Bot
# ─────────────────────────────────────────────────────────────────────────────
aiogram>=3.2.0

# ─────────────────────────────────────────────────────────────────────────────
# HTTP клиенты (шлюзы)
# ─────────────────────────────────────────────────────────────────────────────
httpx>=0.25.2
aiohttp>=3.9.0

# ─────────────────────────────────────────────────────────────────────────────
# SMTP (приём входящих SMS от GoIP)
# ─────────────────────────────────────────────────────────────────────────────
aiosmtpd>=1.4.2

# ─────────────────────────────────────────────────────────────────────────────
# Dev / Test
# ─────────────────────────────────────────────────────────────────────────────
pytest>=7.4.0
pytest-asyncio>=0.21.0
openpyxl>=3.1.0
